+++ lib/src/retrieval/kd_tree/kd_tree_split_strategy.dart
@@ -1,4 +1,4 @@
 enum KDTreeSplitStrategy {
-  widestColumn,
+  largestVariance,
   inOrder,
 }