+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 16.6.3
 - KDTree:
     - `fromIterable` constructor added
+    - `splitStrategy` option added to all constructors
 
 ## 16.6.2
 - KDTree: